+++ bindings/pydrake/planning/planning_py_iris_from_clique_cover.cc
@@ -1,7 +1,7 @@
 #include "drake/bindings/pydrake/documentation_pybind.h"
 #include "drake/bindings/pydrake/pydrake_pybind.h"
 #include "drake/planning/graph_algorithms/max_clique_solver_base.h"
-#include "drake/planning/iris_from_clique_cover.h"
+#include "drake/planning/iris/iris_from_clique_cover.h"
 
 namespace drake {
 namespace pydrake {
