+++ planning/iris/test/iris_from_clique_cover_test.cc
@@ -1,6 +1,6 @@
-#include "drake/planning/iris_from_clique_cover.h"
+#include "drake/planning/iris/iris_from_clique_cover.h"
 
-#include <gtest/gtest.h>
+#include <include/gtest/gtest.h>
 
 #include "drake/common/find_resource.h"
 #include "drake/common/ssize.h"
