+++ tools/install/libdrake/build_components.bzl
@@ -74,8 +74,9 @@ LIBDRAKE_COMPONENTS = [
     "//multibody/triangle_quadrature",
     "//perception",
     "//planning",
-    "//planning/trajectory_optimization",
     "//planning/graph_algorithms",
+    "//planning/iris",
+    "//planning/trajectory_optimization",
     "//solvers",
     "//systems/analysis",
     "//systems/controllers",
